Refactored BoardComponent and GameComponent

+++ src/app/game/game.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, ViewChild, ChangeDetectorRef } from '@angular/core';
+import { Component, OnInit, ViewChild, ChangeDetectorRef, OnDestroy } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 
 import { BoardComponent } from './board/board.component';
@@ -10,7 +10,7 @@ import { AuthService } from '../auth-service.service';
   templateUrl: './game.component.html',
   styleUrls: ['./game.component.css']
 })
-export class GameComponent implements OnInit {
+export class GameComponent implements OnInit, OnDestroy {
   dataLoaded = false;
   showAlert = false;
   alertText: string;
@@ -18,12 +18,12 @@ export class GameComponent implements OnInit {
   private players: any;
   private player: any;
   private opponent: any;
-  private activePlayer: any;
   private points = { red: 0, yellow: 0 };
   private client: any;
   private gameId: string;
   private gameRecord: any;
-  private game: { state: 'in progress' | 'on hold' | 'over' } = { state: 'in progress' };
+  private usersList: any;
+  private gameRecordDestroyed = false;
   @ViewChild(BoardComponent) board: BoardComponent;
 
 
@@ -40,10 +40,10 @@ export class GameComponent implements OnInit {
     this.route.paramMap.subscribe(params => {
       this.gameId = params.get('gameId');
       this.gameRecord = this.client.record.getRecord(this.gameId);
+      this.usersList = this.client.record.getList('users');
       this.gameRecord.subscribe('players', (players: any) => {
         if (players) {
           this.players = players;
-          this.activePlayer = players.red;
           if (this.username === players.red.username) {
             this.player = players.red;
             this.opponent = players.yellow;
@@ -56,75 +56,87 @@ export class GameComponent implements OnInit {
         }
       }, true);
     });
-    this.client.record.getList('users').whenReady((list: any) => {
-      list.on('entry-removed', this.onUserOffline.bind(this));
-    });
-    this.gameRecord.subscribe('moves', this.onMovesUpdate.bind(this), true);
+    this.usersList.on('entry-removed', this.onUserOffline.bind(this));
     this.gameRecord.subscribe('game', this.onGameUpdate.bind(this), true);
     this.gameRecord.subscribe('points', this.onPointsUpdate.bind(this), true);
   }
 
+  ngOnDestroy() {
+    if (!this.gameRecordDestroyed) {
+      this.gameRecord.unsubscribe('game');
+      this.gameRecord.unsubscribe('players');
+    }
+    this.usersList.unsubscribe(this.onUserOffline.bind(this));
+  }
+
   onMove(id: string) {
     this.updateGame(id);
-    this.updateMoves(id);
   }
 
   private updateGame(id: string) {
     if (this.gameover(id)) {
-      const key = this.activePlayer === this.players.red ? 'red' : 'yellow';
+      const key = this.board.activePlayer === this.players.red ? 'red' : 'yellow';
       this.points[`${key}`] += 1;
       this.gameRecord.set('points', this.points);
-      this.gameRecord.set('game', { state: 'over', winner: this.activePlayer });
+      this.gameRecord.set('game.winner', this.board.activePlayer);
+      this.gameRecord.set('game.state', 'over');
     }
-  }
-
-  private updateMoves(id: string) {
-    const moves = this.gameRecord.get('moves') || [];
+    const moves = this.gameRecord.get('game.moves') || [];
     moves.push(id);
-    this.gameRecord.set('moves', moves);
-  }
-
-  private onMovesUpdate(moves: string[] = []) {
-    if (moves.length !== 0) {
-      if (this.board.isEmpty && moves.length > 1) {
-        this.board.replayGame(moves);
-      } else {
-        const id = moves[moves.length - 1];
-        this.board.replayMove(id);
-      }
-      this.activePlayer = moves.length % 2 === 0 ? this.players.red : this.players.yellow;
-      this.board.isEmpty = false;
-    }
+    this.gameRecord.set('game.moves', moves);
   }
 
   private onGameUpdate(game: any) {
     if (game) {
-      this.game = game;
+      const resetGame = this.board.game.state === 'on hold' && game.state === 'in progress';
+      const replayLastMove = this.board.game.moves.length === game.moves.length - 1;
+      this.board.game = game;
+      if (resetGame) {
+        this.resetGame();
+      }
+      if (!this.board.waitingForOpponent) {
+        this.updateBoard(replayLastMove);
+      }
     }
   }
 
-  private onPointsUpdate(points: {red: number, yellow: number}) {
+  private resetGame() {
+    this.board.redMovesFirst = this.board.game.winner.username === this.players.red.username ? false : true;
+    this.board.waitingForOpponent = false;
+    this.board.setActivePlayer(0);
+  }
+
+  private updateBoard(replayLastMove: boolean) {
+    if (replayLastMove) {
+      this.board.replayLastMove();
+    } else {
+      this.board.replayGame();
+    }
+  }
+
+  private onPointsUpdate(points: { red: number, yellow: number }) {
     if (points) {
       this.points = points;
     }
   }
 
   onNewGame() {
-    if (this.game.state === 'on hold') {
-      this.gameRecord.set('moves', []);
-      this.activePlayer = this.username === this.players.red.username ? this.players.red : this.players.yellow;
+    if (this.board.game.state === 'on hold') {
+      this.gameRecord.set('game.moves', []);
       this.gameRecord.set('game.state', 'in progress');
     } else {
       this.gameRecord.set('game.state', 'on hold');
-      this.activePlayer = this.opponent === this.players.red ? this.players.red : this.players.yellow;
     }
   }
 
   onUserOffline(username: string) {
-    this.gameRecord.delete();
-    this.client.record.getRecord(this.username).set('status', 'Online');
-    this.alertText = `${username} has left the game.`;
-    this.showAlert = true;
+    if (this.opponent && username === this.opponent.username) {
+      this.gameRecord.delete();
+      this.client.record.getRecord(this.username).set('status', 'Online');
+      this.alertText = `${username} has left the game.`;
+      this.showAlert = true;
+      this.gameRecordDestroyed = true;
+    }
   }
 
   private gameover(id: string) {
